requirements.txt
fastapi==0.111.0
uvicorn==0.23.2
asyncpg==0.29.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
aiosmtplib==3.0.1
python-jose[cryptography]>=3.0.0